// generated_api/action/model/action.ts
/**
 * Action Service
 * A service that receives incoming notifications and uses pre-defined templates (action objects) to turn those notifications into meaningful actions.
 *
 * OpenAPI spec version: v1beta2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { EmailAction } from './emailAction';
import { WebhookAction } from './webhookAction';

export class Action {
    /**
    * The name of the action. The name is one or more identifier strings separated by dots. Each identifier string consists of lower case letters, digits, and underscores, and cannot start with a digit.
    */
    name: string;
    kind: Action.KindEnum;
    /**
    * Human readable name title for the action. Must be less than 128 characters.
    */
    title: string;
    subject: string;
    body: string;
    addresses: Array<string>;
    /**
    * Only allows the HTTPS scheme. 
    */
    webhookUrl: string;
    /**
    * The (possibly) templated payload body which will be POSTed to the webhookUrl when triggered. 
    */
    webhookPayload: string;

    static discriminator: string | undefined = 'kind';

    static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
        {
            name: 'name',
            baseName: 'name',
            type: 'string',
        },
        {
            name: 'kind',
            baseName: 'kind',
            type: 'Action.KindEnum',
        },
        {
            name: 'title',
            baseName: 'title',
            type: 'string',
        },
        {
            name: 'subject',
            baseName: 'subject',
            type: 'string',
        },
        {
            name: 'body',
            baseName: 'body',
            type: 'string',
        },
        {
            name: 'addresses',
            baseName: 'addresses',
            type: 'Array<string>',
        },
        {
            name: 'webhookUrl',
            baseName: 'webhookUrl',
            type: 'string',
        },
        {
            name: 'webhookPayload',
            baseName: 'webhookPayload',
            type: 'string',
        },
    ];

    static getAttributeTypeMap() {
        return Action.attributeTypeMap;
    }
}

export namespace Action {
    export enum KindEnum {
        Webhook = <any>'webhook',
        Email = <any>'email',
    }
}