// generated_api/action/model/actionCommon.ts
/**
 * Action Service
 * A service that receives incoming notifications and uses pre-defined templates (action objects) to turn those notifications into meaningful actions.
 *
 * OpenAPI spec version: v1beta2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export class ActionCommon {
    /**
     * The name of the action. The name is one or more identifier strings separated by dots. Each identifier string consists of lower case letters, digits, and underscores, and cannot start with a digit.
     */
    public name: string;
    public kind: ActionCommon.KindEnum;
    /**
     * Human readable name title for the action. Must be less than 128 characters.
     */
    public title: string;

    public static discriminator: string | undefined = undefined;

    public static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
        {
            name: 'name',
            baseName: 'name',
            type: 'string',
        },
        {
            name: 'kind',
            baseName: 'kind',
            type: 'ActionCommon.KindEnum',
        },
        {
            name: 'title',
            baseName: 'title',
            type: 'string',
        },
    ];

    public static getAttributeTypeMap() {
        return ActionCommon.attributeTypeMap;
    }
}

export namespace ActionCommon {
    export enum KindEnum {
        Webhook = 'webhook' as any,
        Email = 'email' as any,
    }
}