// generated_api/action/model/actionMutable.d.ts
/**
 * Action Service
 * A service that receives incoming notifications and uses pre-defined templates (action objects) to turn those notifications into meaningful actions.
 *
 * OpenAPI spec version: v1beta2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
export declare class ActionMutable {
    /**
     * Human readable name title for the action. Must be less than 128 characters.
     */
    title: string;
    subject: string;
    body: string;
    addresses: string[];
    /**
     * Only allows the HTTPS scheme.
     */
    webhookUrl: string;
    /**
     * The (possibly) templated payload body which will be POSTed to the webhookUrl when triggered.
     */
    webhookPayload: string;
    static discriminator: string | undefined;
    static attributeTypeMap: Array<{
        name: string;
        baseName: string;
        type: string;
    }>;
    static getAttributeTypeMap(): {
        name: string;
        baseName: string;
        type: string;
    }[];
}