// generated_api/action/model/actionResult.ts
/**
 * Action Service
 * A service that receives incoming notifications and uses pre-defined templates (action objects) to turn those notifications into meaningful actions.
 *
 * OpenAPI spec version: v1beta2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class ActionResult {
    'state'?: ActionResult.StateEnum;
    'statusId'?: string;
    'message'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "state",
            "baseName": "state",
            "type": "ActionResult.StateEnum"
        },
        {
            "name": "statusId",
            "baseName": "statusId",
            "type": "string"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ActionResult.attributeTypeMap;
    }
}

export namespace ActionResult {
    export enum StateEnum {
        QUEUED = <any> 'QUEUED',
        RUNNING = <any> 'RUNNING',
        DONE = <any> 'DONE',
        FAILED = <any> 'FAILED'
    }
}