// generated_api/action/model/deprecatedAction.ts
/**
 * Action Service
 * A service that receives incoming notifications and uses pre-defined templates (action objects) to turn those notifications into meaningful actions.
 *
 * OpenAPI spec version: v1beta2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { DeprecatedEmailAction } from './deprecatedEmailAction';
import { DeprecatedSNSAction } from './deprecatedSNSAction';
import { DeprecatedWebhookAction } from './deprecatedWebhookAction';

export class DeprecatedAction {
    /**
    * The name of the action. the name must be at least 4 alphanumeric characters, and can be segmented with periods.
    */
    'name': string;
    'kind': DeprecatedAction.KindEnum;
    /**
    * Human readable name title for the action. Must be less than 128 characters.
    */
    'title'?: string;
    'subject'?: string;
    'body'?: string;
    'addresses'?: Array<string>;
    'htmlPart'?: string;
    'subjectPart'?: string;
    'textPart'?: string;
    'templateName'?: string;
    /**
    * Only allows the HTTPS scheme. 
    */
    'webhookUrl'?: string;
    'message'?: string;
    'topic'?: string;

    static discriminator: string | undefined = "kind";

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "kind",
            "baseName": "kind",
            "type": "DeprecatedAction.KindEnum"
        },
        {
            "name": "title",
            "baseName": "title",
            "type": "string"
        },
        {
            "name": "subject",
            "baseName": "subject",
            "type": "string"
        },
        {
            "name": "body",
            "baseName": "body",
            "type": "string"
        },
        {
            "name": "addresses",
            "baseName": "addresses",
            "type": "Array<string>"
        },
        {
            "name": "htmlPart",
            "baseName": "htmlPart",
            "type": "string"
        },
        {
            "name": "subjectPart",
            "baseName": "subjectPart",
            "type": "string"
        },
        {
            "name": "textPart",
            "baseName": "textPart",
            "type": "string"
        },
        {
            "name": "templateName",
            "baseName": "templateName",
            "type": "string"
        },
        {
            "name": "webhookUrl",
            "baseName": "webhookUrl",
            "type": "string"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        },
        {
            "name": "topic",
            "baseName": "topic",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return DeprecatedAction.attributeTypeMap;
    }
}

export namespace DeprecatedAction {
    export enum KindEnum {
        Webhook = <any> 'webhook',
        Email = <any> 'email',
        Sns = <any> 'sns'
    }
}