// generated_api/action/model/deprecatedActionCommon.ts
/**
 * Action Service
 * A service that receives incoming notifications and uses pre-defined templates (action objects) to turn those notifications into meaningful actions.
 *
 * OpenAPI spec version: v1beta2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export class DeprecatedActionCommon {
    /**
     * The name of the action. the name must be at least 4 alphanumeric characters, and can be segmented with periods.
     */
    public name: string;
    public kind: DeprecatedActionCommon.KindEnum;
    /**
     * Human readable name title for the action. Must be less than 128 characters.
     */
    public title: string;

    public static discriminator: string | undefined = undefined;

    public static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
        {
            name: 'name',
            baseName: 'name',
            type: 'string',
        },
        {
            name: 'kind',
            baseName: 'kind',
            type: 'DeprecatedActionCommon.KindEnum',
        },
        {
            name: 'title',
            baseName: 'title',
            type: 'string',
        },
    ];

    public static getAttributeTypeMap() {
        return DeprecatedActionCommon.attributeTypeMap;
    }
}

export namespace DeprecatedActionCommon {
    export enum KindEnum {
        Webhook = 'webhook' as any,
        Email = 'email' as any,
        Sns = 'sns' as any,
    }
}