// generated_api/action/model/deprecatedActionMutable.ts
/**
 * Action Service
 * A service that receives incoming notifications and uses pre-defined templates (action objects) to turn those notifications into meaningful actions.
 *
 * OpenAPI spec version: v1beta2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { MutableDeprecatedEmailAction } from './mutableDeprecatedEmailAction';
import { MutableDeprecatedSNSAction } from './mutableDeprecatedSNSAction';
import { MutableDeprecatedWebhookAction } from './mutableDeprecatedWebhookAction';

export class DeprecatedActionMutable {
    /**
    * Human readable name title for the action. Must be less than 128 characters.
    */
    'title'?: string;
    'subject'?: string;
    'body'?: string;
    'addresses'?: Array<string>;
    'htmlPart'?: string;
    'subjectPart'?: string;
    'textPart'?: string;
    'templateName'?: string;
    /**
    * Only allows the HTTPS scheme. 
    */
    'webhookUrl'?: string;
    'message'?: string;
    'topic'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "title",
            "baseName": "title",
            "type": "string"
        },
        {
            "name": "subject",
            "baseName": "subject",
            "type": "string"
        },
        {
            "name": "body",
            "baseName": "body",
            "type": "string"
        },
        {
            "name": "addresses",
            "baseName": "addresses",
            "type": "Array<string>"
        },
        {
            "name": "htmlPart",
            "baseName": "htmlPart",
            "type": "string"
        },
        {
            "name": "subjectPart",
            "baseName": "subjectPart",
            "type": "string"
        },
        {
            "name": "textPart",
            "baseName": "textPart",
            "type": "string"
        },
        {
            "name": "templateName",
            "baseName": "templateName",
            "type": "string"
        },
        {
            "name": "webhookUrl",
            "baseName": "webhookUrl",
            "type": "string"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        },
        {
            "name": "topic",
            "baseName": "topic",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return DeprecatedActionMutable.attributeTypeMap;
    }
}