// generated_api/action/model/deprecatedActionMutableCommon.ts
/**
 * Action Service
 * A service that receives incoming notifications and uses pre-defined templates (action objects) to turn those notifications into meaningful actions.
 *
 * OpenAPI spec version: v1beta2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export class DeprecatedActionMutableCommon {
    /**
     * Human readable name title for the action. Must be less than 128 characters.
     */
    public title: string;

    public static discriminator: string | undefined = undefined;

    public static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
        {
            name: 'title',
            baseName: 'title',
            type: 'string',
        },
    ];

    public static getAttributeTypeMap() {
        return DeprecatedActionMutableCommon.attributeTypeMap;
    }
}