// generated_api/action/model/deprecatedEmailAction.ts
/**
 * Action Service
 * A service that receives incoming notifications and uses pre-defined templates (action objects) to turn those notifications into meaningful actions.
 *
 * OpenAPI spec version: v1beta2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { DeprecatedActionCommon } from './deprecatedActionCommon';
import { MutableDeprecatedEmailAction } from './mutableDeprecatedEmailAction';

export class DeprecatedEmailAction {
    /**
     * The name of the action. the name must be at least 4 alphanumeric characters, and can be segmented with periods.
     */
    public name: string;
    public kind: DeprecatedEmailAction.KindEnum;
    /**
     * Human readable name title for the action. Must be less than 128 characters.
     */
    public title: string;
    public subject: string;
    public body: string;
    public addresses: string[];
    public htmlPart: string;
    public subjectPart: string;
    public textPart: string;
    public templateName: string;

    public static discriminator: string | undefined = undefined;

    public static attributeTypeMap: Array<{ name: string; baseName: string; type: string }> = [
        {
            name: 'name',
            baseName: 'name',
            type: 'string',
        },
        {
            name: 'kind',
            baseName: 'kind',
            type: 'DeprecatedEmailAction.KindEnum',
        },
        {
            name: 'title',
            baseName: 'title',
            type: 'string',
        },
        {
            name: 'subject',
            baseName: 'subject',
            type: 'string',
        },
        {
            name: 'body',
            baseName: 'body',
            type: 'string',
        },
        {
            name: 'addresses',
            baseName: 'addresses',
            type: 'Array<string>',
        },
        {
            name: 'htmlPart',
            baseName: 'htmlPart',
            type: 'string',
        },
        {
            name: 'subjectPart',
            baseName: 'subjectPart',
            type: 'string',
        },
        {
            name: 'textPart',
            baseName: 'textPart',
            type: 'string',
        },
        {
            name: 'templateName',
            baseName: 'templateName',
            type: 'string',
        },
    ];

    public static getAttributeTypeMap() {
        return DeprecatedEmailAction.attributeTypeMap;
    }
}

export namespace DeprecatedEmailAction {
    export enum KindEnum {
        Webhook = 'webhook' as any,
        Email = 'email' as any,
        Sns = 'sns' as any,
    }
}