// generated_api/action/model/deprecatedSNSAction.ts
/**
 * Action Service
 * A service that receives incoming notifications and uses pre-defined templates (action objects) to turn those notifications into meaningful actions.
 *
 * OpenAPI spec version: v1beta2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { DeprecatedActionCommon } from './deprecatedActionCommon';
import { MutableDeprecatedSNSAction } from './mutableDeprecatedSNSAction';

export class DeprecatedSNSAction {
    /**
    * The name of the action. the name must be at least 4 alphanumeric characters, and can be segmented with periods.
    */
    'name': string;
    'kind': DeprecatedSNSAction.KindEnum;
    /**
    * Human readable name title for the action. Must be less than 128 characters.
    */
    'title'?: string;
    'topic'?: string;
    'message'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "kind",
            "baseName": "kind",
            "type": "DeprecatedSNSAction.KindEnum"
        },
        {
            "name": "title",
            "baseName": "title",
            "type": "string"
        },
        {
            "name": "topic",
            "baseName": "topic",
            "type": "string"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return DeprecatedSNSAction.attributeTypeMap;
    }
}

export namespace DeprecatedSNSAction {
    export enum KindEnum {
        Webhook = <any> 'webhook',
        Email = <any> 'email',
        Sns = <any> 'sns'
    }
}